test_tabular: Pull out some auxiliary code
These are only used in test_tabular.py, so this is mostly for
organizational clarity, though some of these may be helpful to
libraries that use pyout.
